setup controller,view,model for one to one relationship

+++ app/Http/Controllers/AdminController.php
@@ -5,11 +5,14 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use DB;
 use Hash;
+use App\Models\User;
+use App\Models\Identitycard;
 
 class AdminController extends Controller
 {
     public function index() {
-        $users = DB::select('select * from users where user_type = "member"');
+        // $users = DB::select('select * from users where user_type = "member"');
+        $users = User::all()->where('user_type','member');
         return view('admin.index',['users'=>$users]);
     }
 
@@ -18,8 +21,13 @@ class AdminController extends Controller
     public function createPage() {
         return view('admin.create');
     }
+
+
     // Create User
     public function create(Request $request) {
+
+        
+
         $name = $request->input('UserName');
         $email = $request->input('UserEmail');
         $password = Hash::make($request->input('UserPassword'));
@@ -34,31 +42,43 @@ class AdminController extends Controller
             $picture->move(public_path('public/Image'), $filename);
         }
 
-        DB::insert("INSERT INTO `users`(`name`, `email`,`profile_picture`, `password`) VALUES (?,?,?,?)",
-        [$request->input('UserName'),$request->input('UserEmail'),$filename,$password]);
 
-        return redirect('/admin');
+        $user = new User;
+        $user->name = $request->input('UserName');
+        $user->email = $request->input('UserEmail');
+        $user->profile_picture = $filename;
+        $user->password = Hash::make($request->input('UserPassword'));
+        $user->save();
+
+
+        $identity = new Identitycard;
+        $identity->user_id = $user->id;
+        $identity->identity_number = rand(1000000,9999999);
+        $identity->phone_number = rand(100000000,999999999);
+        $identity->save();
+
+      return redirect('/admin');
     }
 
     // View User
     public function show($id) {
-        $user = DB::select('select *  from users where id = ?',[$id]);
+        $user = DB::select('select * from users INNER JOIN identitycards ON users.id = identitycards.user_id where users.id = ?',[$id]);
         return view('admin.show',['user'=>$user]);
     }
 
     // Edit User Page
     public function editPage($id) {
-        $user = DB::select('select *  from users where id = ?',[$id]);
+        $user = DB::select('select * from users INNER JOIN identitycards ON users.id = identitycards.user_id where users.id = ?',[$id]);
         return view('admin.edit',['user'=>$user]);
     }
 
     // edit User
     public function edit(Request $request,$id) {
         $filename = "";
-        if($request->file('UserProfile')){
+        if($request->file('UserProfile') ){
             // remove existing file
             $image = DB::select('select profile_picture from users where id = ?',[$id]);
-            if(file_exists(public_path().'/public/Image/'.$image[0]->profile_picture)){
+            if(file_exists(public_path().'/public/Image/'.$image[0]->profile_picture) && $image[0]->profile_picture){
                 unlink(public_path().'/public/Image/'.$image[0]->profile_picture);
             }
             
@@ -67,12 +87,21 @@ class AdminController extends Controller
             $file-> move(public_path('public/Image'), $filename);
         }
         
-        if($filename == "")
+        if($filename == "") {
             DB::update("UPDATE `users` SET `name`=?  WHERE id = ?",
             [$request->input('UserName'),$id]);
-        else
+
+            DB::update("UPDATE `identitycards` SET `phone_number`=?  WHERE user_id = ?",
+            [$request->input('UserPhone'),$id]);
+        }
+
+        else {
             DB::update("UPDATE `users` SET `name`=?, `profile_picture`= ? WHERE id = ?",
             [$request->input('UserName'),$filename,$id]);
+
+            DB::update("UPDATE `identitycards` SET `phone_number`=?  WHERE user_id = ?",
+            [$request->input('UserPhone'),$id]);
+        }
         
         return redirect('/admin');
     }
@@ -80,12 +109,13 @@ class AdminController extends Controller
     // Delete User
     public function delete($id) {
         $image = DB::select('select profile_picture from users where id = ?',[$id]);
-        
+        echo $id;
         if($image[0]->profile_picture) {
             if(file_exists(public_path().'/public/Image/'.$image[0]->profile_picture)){
                 unlink(public_path().'/public/Image/'.$image[0]->profile_picture);
             }
-            DB::delete('delete from users where id = ?',[$id]);
+            // DB::delete('delete from users where users.id = ?',[$id]);
+            DB::table('users')->where('id', $id)->delete();
             
         }
         return redirect('/admin');
